+++ src/views/assets/Asset/AssetList/components/AssetBulkUpdateDialog.vue
