+++ src/views/assets/Asset/AssetList/components/NodeAssetsUpdateDialog.vue
