+++ src/views/assets/Asset/AssetList/index.vue
@@ -1,306 +1,66 @@
 <template>
-  <div>
-    <TabPage :active-menu.sync="tab.activeMenu" :submenu="tab.submenu" @tab-click="handleTabChange">
-      <TreeTable
-        v-if="show"
-        ref="TreeList"
-        :table-config="tableConfig"
-        :help-message="helpMessage"
-        :header-actions="headerActions"
-        :tree-setting="treeSetting"
-        :show-tree="showTree"
-      >
-        <TreeMenu
-          slot="rMenu"
-          :tree="treeRef"
-          @showAll="showAll"
-        />
-      </TreeTable>
-    </TabPage>
-    <AssetBulkUpdateDialog
-      :visible.sync="updateSelectedDialogSetting.visible"
-      v-bind="updateSelectedDialogSetting"
-    />
-    <PlatformDialog :visible.sync="showPlatform" :category="category" />
-  </div>
+  <TabPage :active-menu.sync="config.activeMenu" :submenu="config.submenu" />
 </template>
 
 <script>
 import { TabPage } from '@/layout/components'
-import { TreeTable } from '@/components'
-import $ from '@/utils/jquery-vendor'
-import { mapGetters } from 'vuex'
-import { connectivityMeta } from '@/components/AccountListTable/const'
-import AssetBulkUpdateDialog from './AssetBulkUpdateDialog'
-import PlatformDialog from './PlatformDialog'
-import TreeMenu from './TreeMenu'
-import { Categories } from '@/views/assets/const'
-import {
-  DetailFormatter,
-  ActionsFormatter,
-  TagsFormatter,
-  NestedObjectFormatter,
-  ChoicesDisplayFormatter
-} from '@/components/TableFormatters'
 
 export default {
+  name: 'Index',
   components: {
-    AssetBulkUpdateDialog,
-    TreeMenu,
-    PlatformDialog,
-    TreeTable,
     TabPage
   },
   data() {
-    const vm = this
     return {
-      treeRef: null,
-      showPlatform: false,
-      showTree: true,
-      show: true,
-      tab: {
+      config: {
+        activeMenu: 'all',
         submenu: [
-          { name: 'all', title: '所有' },
-          ...Categories
-        ],
-        activeMenu: 'all'
-      },
-      category: '',
-      treeSetting: {
-        showMenu: true,
-        showRefresh: true,
-        showAssets: false,
-        showCreate: true,
-        showUpdate: true,
-        showDelete: true,
-        hasRightMenu: true,
-        url: '/api/v1/assets/assets/',
-        nodeUrl: '/api/v1/assets/nodes/',
-        // ?assets=0不显示资产. =1显示资产
-        treeUrl: '/api/v1/assets/nodes/children/tree/?assets=0'
-      },
-      tableConfig: {
-        url: '/api/v1/assets/assets/',
-        hasTree: true,
-        columns: [
-          'name', 'ip', 'category', 'type', 'platform',
-          'labels', 'nodes', 'is_active', 'connectivity',
-          'created_by', 'date_created', 'comment', 'org_name',
-          'actions'
-        ],
-        columnsShow: {
-          min: ['name', 'ip', 'actions'],
-          default: [
-            'name', 'ip', 'platform', 'category', 'type',
-            'connectivity', 'actions'
-          ]
-        },
-        columnsMeta: {
-          type: { formatter: ChoicesDisplayFormatter },
-          category: { formatter: ChoicesDisplayFormatter },
-          name: {
-            formatter: DetailFormatter,
-            formatterArgs: {
-              route: 'AssetDetail'
-            },
-            showOverflowTooltip: true,
-            sortable: true
-          },
-          ip: { sortable: 'custom', width: '140px' },
-          comment: { showOverflowTooltip: true },
-          connectivity: connectivityMeta,
-          labels_display: { formatter: TagsFormatter },
-          nodes: { formatter: NestedObjectFormatter },
-          actions: {
-            formatter: ActionsFormatter,
-            formatterArgs: {
-              performDelete: ({ row, col }) => {
-                const id = row.id
-                const url = `/api/v1/assets/assets/${id}/`
-                return this.$axios.delete(url)
-              },
-              extraActions: [
-                {
-                  name: 'View',
-                  title: this.$t(`common.UpdateAssetDetail`),
-                  type: 'primary',
-                  can: vm.$hasPerm('assets.refresh_assethardwareinfo'),
-                  callback: function({ cellValue, tableData, row }) {
-                    return this.$router.push({
-                      name: 'AssetMoreInformationEdit',
-                      params: { id: row.id }
-                    })
-                  }
-                }
-              ]
-            }
-          }
-        }
-      },
-      headerActions: {
-        createRoute: () => {
-          return {
-            name: 'AssetCreate',
-            query: this.$route.query
-          }
-        },
-        onCreate: () => {
-          this.showPlatform = true
-        },
-        createInNewPage: true,
-        searchConfig: {
-          options: [
-            { label: this.$t('assets.Label'), value: 'label' }
-          ]
-        },
-        extraActions: [
           {
-            name: this.$t('xpack.Cloud.CloudSync'),
-            title: this.$t('xpack.Cloud.CloudSync'),
-            has: () => vm.$hasPerm('xpack.view_account') && vm.$hasLicense(),
-            callback: () => this.$router.push({ name: 'CloudCenter' })
-          }
-        ],
-        extraMoreActions: [
-          {
-            name: 'DeactiveSelected',
-            title: this.$t('assets.DeactiveSelected'),
-            type: 'primary',
-            can: ({ selectedRows }) => {
-              return selectedRows.length > 0 && vm.$hasPerm('assets.change_asset')
-            },
-            callback: function({ selectedRows }) {
-              const ids = selectedRows.map((v) => {
-                return { pk: v.id, is_active: false }
-              })
-              this.$axios.patch(`/api/v1/assets/assets/`, ids).then(res => {
-                this.$message.success(this.$t('common.updateSuccessMsg'))
-              }).catch(err => {
-                this.$message.error(this.$t('common.updateErrorMsg' + ' ' + err))
-              })
-            }.bind(this)
+            title: this.$t('assets.All'),
+            name: 'all',
+            component: () => import('@/views/assets/Asset/AssetList/AllList.vue')
           },
           {
-            name: 'ActiveSelected',
-            title: this.$t('assets.ActiveSelected'),
-            type: 'primary',
-            can: ({ selectedRows }) => {
-              return selectedRows.length > 0 && vm.$hasPerm('assets.change_asset')
-            },
-            callback: function({ selectedRows }) {
-              const ids = selectedRows.map((v) => {
-                return { pk: v.id, is_active: true }
-              })
-              this.$axios.patch(`/api/v1/assets/assets/`, ids).then(res => {
-                this.$message.success(this.$t('common.updateSuccessMsg'))
-              }).catch(err => {
-                this.$message.error(this.$t('common.updateErrorMsg' + ' ' + err))
-              })
-            }.bind(this)
+            title: this.$t('applications.host'),
+            name: 'host',
+            hidden: () => false,
+            component: () => import('@/views/assets/Asset/AssetList/HostList.vue')
           },
           {
-            name: 'actionUpdateSelected',
-            title: this.$t('common.updateSelected'),
-            can: ({ selectedRows }) => {
-              return selectedRows.length > 0 &&
-                !vm.currentOrgIsRoot &&
-                vm.$hasPerm('assets.change_asset')
-            },
-            callback: ({ selectedRows }) => {
-              vm.updateSelectedDialogSetting.selectedRows = selectedRows
-              vm.updateSelectedDialogSetting.visible = true
-            }
+            title: this.$t('route.NetworkDevices'),
+            name: 'networking',
+            hidden: () => false,
+            component: () => import('@/views/assets/Asset/AssetList/NetworkList.vue')
           },
           {
-            name: 'RemoveFromCurrentNode',
-            title: this.$t('assets.RemoveFromCurrentNode'),
-            can: ({ selectedRows }) => {
-              if (!vm.$route.query.node) {
-                return false
-              }
-              return selectedRows.length > 0 &&
-                !vm.currentOrgIsRoot &&
-                vm.$hasPerm('assets.change_node')
-            },
-            callback: function({ selectedRows, reloadTable }) {
-              const assetsId = []
-              for (const item of selectedRows) {
-                assetsId.push(item.id)
-              }
-              const nodeId = this.$route.query.node
-              if (!nodeId) {
-                return
-              }
-              const url = `/api/v1/assets/nodes/${nodeId}/assets/remove/`
-              this.$axios.put(url, { assets: assetsId }).then(res => {
-                this.$message.success(this.$t('common.removeSuccessMsg'))
-                reloadTable()
-              }).catch(err => {
-                this.$message.error(this.$t('common.removeErrorMsg' + ' ' + err))
-              })
-            }.bind(this)
+            title: this.$t('route.Databases'),
+            name: 'database',
+            hidden: () => false,
+            component: () => import('@/views/assets/Asset/AssetList/DatabaseList.vue')
+          },
+          {
+            title: this.$t('assets.CloudPlatform'),
+            name: 'cloud',
+            hidden: () => false,
+            component: () => import('@/views/assets/Asset/AssetList/CloudsPlatformList.vue')
+          },
+          {
+            title: 'Web',
+            name: 'web',
+            hidden: () => false,
+            component: () => import('@/views/assets/Asset/AssetList/WebList.vue')
           }
         ]
-      },
-      helpMessage: this.$t('assets.AssetListHelpMessage'),
-      updateSelectedDialogSetting: {
-        visible: false,
-        selectedRows: []
       }
     }
   },
   computed: {
-    ...mapGetters(['currentOrgIsRoot'])
   },
   mounted() {
-    this.decorateRMenu()
-    this.treeSetting.hasRightMenu = !this.currentOrgIsRoot
-    this.treeSetting.showCreate = this.$hasPerm('assets.add_node')
-    this.treeSetting.showUpdate = this.$hasPerm('assets.change_node')
-    this.treeSetting.showDelete = this.$hasPerm('assets.delete_node')
-    this.treeRef = this.$refs.TreeList
-    this.handleTabChange()
-  },
-  methods: {
-    decorateRMenu() {
-      const show_current_asset = this.$cookie.get('show_current_asset') || '0'
-      if (show_current_asset === '1') {
-        $('#m_show_asset_all_children_node').css('color', '#606266')
-        $('#m_show_asset_only_current_node').css('color', 'green')
-      } else {
-        $('#m_show_asset_all_children_node').css('color', 'green')
-        $('#m_show_asset_only_current_node').css('color', '#606266')
-      }
-    },
-    showAll({ node, showCurrentAsset }) {
-      this.$cookie.set('show_current_asset', showCurrentAsset, 1)
-      this.decorateRMenu()
-      const url = `${this.treeSetting.url}?node_id=${node.meta.data.id}&show_current_asset=${showCurrentAsset}`
-      this.$refs.TreeList.$refs.TreeTable.handleUrlChange(url)
-    },
-    handleTabChange(item) {
-      const category = item ? item.name : this.tab.activeMenu
-      this.category = category
-      console.log('category: ', category)
-      this.show = false
-      setTimeout(() => {
-        let url = '/api/v1/assets/assets/'
-        const showTree = category === 'all'
-        if (category !== 'all') {
-          url = `${url}?category=${category}`
-        }
-        this.treeSetting.url = url
-        this.showTree = showTree
-        this.tableConfig.url = url
-        this.show = true
-      }, 100)
-    }
   }
 }
 </script>
 
-<style lang="scss" scoped>
-.asset-select-dialog >>> .transition-box:first-child {
-  background-color: #f3f3f3;
-}
+<style scoped>
+
 </style>
